backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pillow>=10.4.0
pydantic>=2.8.0
PyYAML>=6.0.2
torch>=2.5.0
torchvision>=0.20.0
numpy>=2.0.0
opencv-python>=4.10.0
onnxruntime>=1.20.0
huggingface_hub>=0.26.0
gdown>=5.2.0
timm>=1.0.20
thop>=0.1.1
omegaconf>=2.3.0
# STDN runs in a separate TensorFlow-compatible runtime.
